# blockchain/standalone: Add subsidy calc funcs.