**This requires PR #1809**.

This adds a subsidy cache with funcs to calculate subsidies for proof-of-work, votes, and the treasury to the `blockchain/standalone` module.  Since the existing functions in `blockchain` are easy to use     improperly and have poor performance under certain access patterns, rather than copying them over, this provide new implementations of better optimized functions with less potential cases for misuse.

Note that because subsidy generation is consensus critical, great care has been taken to ensure the same semantics are retained under the scenarios which the consensus code invokes the functions, although the semantics of the functions themselves have changed when compared to their existing counterparts when called in circumstances that the consensus code avoids as discussed below.

The existing subsidy code in blockchain assumes that it will never be called in circumstances which would violate consensus such as with negative block heights and when there are less than the minimum required number of votes and consequently produces invalid subsidy values in those types of cases.

This does not pose a problem for the existing consensus code itself because it does, in fact, correctly avoid calling the functions in scenarios that would produce incorrect results.  However, since these functions are exported for external use, forcing consumers to only call the functions under the (previously undocumented) correct scenarios is highly error prone.

Consequently, these new functions produce results that match the consensus behavior even if called under those circumstances which are currently avoided.

The following is an overview of the changes to the results of the functions as compared to their existing counterparts:

- Negative block heights now returns 0 subsidy in all cases
- Height 0 now returns 0 subsidy in all cases
- Height 1 now returns the full amount for the PoW subsidy and 0 for the vote and treasury subsidies as compared to reduced values that the actual consensus rules do not use
- Heights prior to one block before the point at which voting begins now return 0 vote subsidy as compared to reduced values that the actual consensus rules ignore at those heights
- Specifying less than the minimum number of votes required for a block to be considered valid by consensus along with heights at or after the point voting begins now returns 0 PoW and Treasury subsidies as compared to further reduced values that the actual consensus rules ignore due to the block being invalid

A few other improvements include:

- All subsidy calculation funcs are now methods on the cache instance which means it is no longer possible to corrupt the cache by calling the functions with different network parameters
- The cache can no longer be made to grow indefinitely
  - Note that this is not an exploit in the existing code since the internal code is very restrictive in how it uses the functions as previously mentioned
- Improves future scalability by avoiding recalculations on all intervals after the max potential block subsidy has been fully reduced to 0
- Improve sparse access efficiency to O(log N) vs O(N)

Finally, it also adds a benchmark, updates the documentation, and includes comprehensive tests.

The following is a before and after comparison of calculating subsidies for various sparse intervals:

```
benchmark                         old ns/op    new ns/op    delta
-------------------------------------------------------------------
BenchmarkCalcSubsidyCacheSparse   56766        17243        -69.62%

benchmark                         old allocs   new allocs   delta
-------------------------------------------------------------------
BenchmarkCalcSubsidyCacheSparse   6            5            -16.67%

benchmark                         old bytes    new bytes    delta
-------------------------------------------------------------------
BenchmarkCalcSubsidyCacheSparse   1163         416          -64.23%
```