@davecgh 2019-08-05
$ go test -coverprofile=cov.out && go tool cover -func=cov.out
PASS
coverage: 100.0% of statements
ok      github.com/decred/dcrd/blockchain/standalone    0.513s
github.com/decred/dcrd/blockchain/standalone/error.go:36:       String                  100.0%
github.com/decred/dcrd/blockchain/standalone/error.go:52:       Error                   100.0%
github.com/decred/dcrd/blockchain/standalone/error.go:57:       ruleError               100.0%
github.com/decred/dcrd/blockchain/standalone/error.go:63:       IsErrorCode             100.0%
github.com/decred/dcrd/blockchain/standalone/merkle.go:29:      CalcMerkleRootInPlace   100.0%
github.com/decred/dcrd/blockchain/standalone/merkle.go:91:      CalcMerkleRoot          100.0%
github.com/decred/dcrd/blockchain/standalone/merkle.go:112:     CalcTxTreeMerkleRoot    100.0%
github.com/decred/dcrd/blockchain/standalone/pow.go:27:         HashToBig               100.0%
github.com/decred/dcrd/blockchain/standalone/pow.go:62:         CompactToBig            100.0%
github.com/decred/dcrd/blockchain/standalone/pow.go:94:         BigToCompact            100.0%
github.com/decred/dcrd/blockchain/standalone/pow.go:142:        CalcWork                100.0%
github.com/decred/dcrd/blockchain/standalone/pow.go:158:        checkProofOfWorkRange   100.0%
github.com/decred/dcrd/blockchain/standalone/pow.go:177:        CheckProofOfWorkRange   100.0%
github.com/decred/dcrd/blockchain/standalone/pow.go:185:        CheckProofOfWork        100.0%
github.com/decred/dcrd/blockchain/standalone/subsidy.go:118:    NewSubsidyCache         100.0%
github.com/decred/dcrd/blockchain/standalone/subsidy.go:140:    Len                     100.0%
github.com/decred/dcrd/blockchain/standalone/subsidy.go:141:    Less                    100.0%
github.com/decred/dcrd/blockchain/standalone/subsidy.go:142:    Swap                    100.0%
github.com/decred/dcrd/blockchain/standalone/subsidy.go:157:    CalcBlockSubsidy        100.0%
github.com/decred/dcrd/blockchain/standalone/subsidy.go:249:    CalcWorkSubsidy         100.0%
github.com/decred/dcrd/blockchain/standalone/subsidy.go:290:    CalcStakeVoteSubsidy    100.0%
github.com/decred/dcrd/blockchain/standalone/subsidy.go:317:    CalcTreasurySubsidy     100.0%
total:                                                          (statements)            100.0%